+++ HealthMonitor.Domain/Models/User/UserInfoDto.cs
@@ -10,7 +10,7 @@ namespace HealthMonitor.Domain.Models.User
         public string? Gender { get; set; }
         public int? Age { get; set; }
         public int? Height { get; set; }
-        public int? Weight { get; set; }
+        public float? Weight { get; set; }
         public string? Goal { get; set; }
         public string? Bio { get; set; }
         public bool OnboardingCompleted { get; set; } = false;
